Fix int conversion warning.

+++ src/mip/mip_result.h
@@ -98,7 +98,7 @@ struct CmdResult
     CmdResult& operator=(const CmdResult& other) = default;
     CmdResult& operator=(C::mip_cmd_result other) { value = other; return *this; }
 
-    static constexpr CmdResult userResult(uint32_t n) { return CmdResult(static_cast<C::mip_cmd_result>(STATUS_USER - n)); }
+    static constexpr CmdResult userResult(uint8_t n) { return CmdResult(static_cast<C::mip_cmd_result>(STATUS_USER - int8_t(n))); }
     static constexpr CmdResult fromAckNack(uint8_t code) { return CmdResult(static_cast<C::mip_cmd_result>(code)); }
 
     operator const void*() const { return isAck() ? this : nullptr; }
